feat: 노트 목록 페이지 동적 라우팅([page]) 전환 (#15)
- /notes → /notes/1 리다이렉트로 변경
- app/notes/[page]/page.tsx: 카테고리·정렬·페이지네이션 통합 목록 페이지 신규 추가
- layout.tsx: main에 flex flex-col 추가로 Pagination 하단 고정 레이아웃 지원
