+++ app/notes/page.tsx
@@ -1,50 +1,11 @@
-import { CategoryFilter } from "@/shared/components/CategoryFilter";
-import { PageHeader } from "@/shared/components/PageHeader";
-import { SortSelect } from "@/shared/components/SortSelect";
-import {
-  CATEGORIES,
-  isCategory,
-  isCategorySortValue,
-  SORT_OPTIONS,
-} from "@/shared/constants/category";
+import { redirect } from "next/navigation";
 
-type NotesPageProps = {
-  searchParams: Promise<{
-    category?: string | string[];
-    sort?: string | string[];
-  }>;
-};
-
-export default async function NotesPage({ searchParams }: NotesPageProps) {
-  const { category, sort } = await searchParams;
-
-  // 배열로 전달된 경우 첫 번째 값만 사용
-  const rawCategory = Array.isArray(category) ? category[0] : category;
-  const rawSort = Array.isArray(sort) ? sort[0] : sort;
-
-  // 유효하지 않은 값은 기본값으로 폴백
-  const selectedSort = isCategorySortValue(rawSort)
-    ? rawSort
-    : SORT_OPTIONS[0].value;
-  const selectedCategory = isCategory(rawCategory)
-    ? rawCategory
-    : CATEGORIES[0].value;
-
-  return (
-    <>
-      <PageHeader className="flex flex-col gap-4">
-        <div className="flex items-center justify-between gap-4">
-          <PageHeader.Title>Notes</PageHeader.Title>
-          <PageHeader.NewButton href="/notes/new">
-            + New Note
-          </PageHeader.NewButton>
-        </div>
-
-        <div className="flex items-center justify-between gap-4">
-          <CategoryFilter selected={selectedCategory} />
-          <SortSelect selected={selectedSort} />
-        </div>
-      </PageHeader>
-    </>
-  );
+/**
+ * /notes 기본 경로 → /notes/1 로 리다이렉트
+ *
+ * - 페이지네이션은 path param([page]) 기반으로 동작
+ * - searchParams(category, sort)는 없으면 [page] 페이지에서 기본값으로 처리
+ */
+export default async function NotesIndexPage() {
+  redirect("/notes/1");
 }
